Make sure the ALIGN() macro uses u_long on all platforms for consistency, and update the comment block accordingly.

+++ sys/arch/arm/include/param.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: param.h,v 1.14 2010/11/20 20:15:28 miod Exp $	*/
+/*	$OpenBSD: param.h,v 1.15 2011/04/07 15:45:16 miod Exp $	*/
 /*	$NetBSD: param.h,v 1.9 2002/03/24 03:37:23 thorpej Exp $	*/
 
 /*
@@ -138,11 +138,11 @@ void	delay (unsigned);
  *
  */
 #define ALIGNBYTES		(sizeof(int) - 1)
-#define ALIGN(p)		(((u_int)(p) + ALIGNBYTES) &~ ALIGNBYTES)
+#define ALIGN(p)		(((u_long)(p) + ALIGNBYTES) &~ ALIGNBYTES)
 #define ALIGNED_POINTER(p,t)	((((u_long)(p)) & (sizeof(t)-1)) == 0)
 /* ARM-specific macro to align a stack pointer (downwards). */
 #define STACKALIGNBYTES		(8 - 1)
-#define STACKALIGN(p)		((u_int)(p) &~ STACKALIGNBYTES)
+#define STACKALIGN(p)		((u_long)(p) &~ STACKALIGNBYTES)
 
 #define	DEV_BSHIFT	9		/* log2(DEV_BSIZE) */
 #define	DEV_BSIZE	(1 << DEV_BSHIFT)
