+++ sys/arch/vax/include/param.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: param.h,v 1.35 2011/03/23 16:54:37 pirofti Exp $ */
+/*	$OpenBSD: param.h,v 1.36 2011/04/07 15:45:18 miod Exp $ */
 /*      $NetBSD: param.h,v 1.39 1999/10/22 21:14:34 ragge Exp $    */
 /*-
  * Copyright (c) 1990 The Regents of the University of California.
@@ -49,7 +49,7 @@
 
 /*
  * Round p (pointer or byte index) up to a correctly-aligned value
- * for all data types (int, long, ...).   The result is u_int and
+ * for all data types (int, long, ...).   The result is u_long and
  * must be cast to any desired pointer type.
  *
  * ALIGNED_POINTER is a boolean macro that checks whether an address
@@ -60,7 +60,7 @@
  */
 
 #define ALIGNBYTES		(sizeof(int) - 1)
-#define ALIGN(p)		(((u_int)(p) + ALIGNBYTES) &~ ALIGNBYTES)
+#define ALIGN(p)		(((u_long)(p) + ALIGNBYTES) &~ ALIGNBYTES)
 #define ALIGNED_POINTER(p,t)	((((u_long)(p)) & (sizeof(t)-1)) == 0)
 
 #define	PGSHIFT		12			/* LOG2(NBPG) */
